infra/cdk/pyproject.toml
[project]
name = "llm-finetune-infra"
version = "1.0.0"
requires-python = ">=3.10,<3.12"
dependencies = [
  "aws-cdk-lib>=2.150.0,<3.0.0",
  "constructs>=10.3.0,<11.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -v"
